// lib/prediction-market/kalshi/dflow/prediction-markets/types.ts
export type MarketStatus =
  | "initialized"
  | "active"
  | "inactive"
  | "closed"
  | "determined";